==> Scripts/15_obj1_lasso_results.R <==
### Summmer Project -- Objective 1 : Multivariate - LASSO stability selection on metabolites and air pollutants
### 31st May 2021 edited on 16th Jun- Alicia 

## Q: Which metabolites are statistically correlated with which exposures in the data?
## --> Multivariate LASSO regression on each exposure 

# set path to include my libraries
.libPaths(c("/rds/general/user/cls1017/home/R/x86_64-redhat-linux-gnu-library/3.6",
            "/usr/lib64/R/library",
            "/usr/share/R/library"))

# devtools::install("../Functions/focus-master")

LoadPackages=function(packages){
  for (i in 1:length(packages)){
    suppressPackageStartupMessages(library(packages[i], character.only=TRUE))
  }
}
LoadPackages(c("pheatmap","corpcor","abind","parallel",
               "RColorBrewer","ppcor","mvtnorm",
               "pROC","stabs","huge","pulsar",
               "QUIC","igraph","glasso","glassoFast",
               "colorspace","glmnet","dplyr", "devtools", "focus"))


path <- "/rds/general/project/hda_students_data/live/Group2/Alicia/Summer_project/"

## -----------------------------------------------------------------------------------------------------------------------
group1_med = F
met_no_bio = T
## -----------------------------------------------------------------------------------------------------------------------
if (group1_med){
  metab_clus=readRDS(paste(path, "Results_report/cluster_manual/denoised_group1_med_base.rds", sep=""))
  metab_clus = metab_clus[order(as.numeric(row.names(metab_clus))), ] %>% as.matrix()
  combine = readRDS(paste(path, "Results_report/cluster_manual/combined_group1_med_base.rds", sep =""))
  combine = combine[order(as.numeric(row.names(combine))),] %>% as.matrix()
  path_to_Results_report = "/rds/general/project/hda_students_data/live/Group2/Alicia/Summer_project/Results_report/Multivariate_obj1/cluster_manual/group1_med/"
  path_to_save = "/rds/general/project/hda_students_data/live/Group2/Alicia/Summer_project/Figure_report/Multivariate_obj1/cluster_manual/group1_med/"
}

if (group1_cen){
  metab_clus=readRDS(paste(path, "Results_report/cluster_manual/denoised_group1_con_base.rds", sep=""))
  metab_clus = metab_clus[order(as.numeric(row.names(metab_clus))), ] %>% as.matrix()
  combine = readRDS(paste(path, "Results_report/cluster_manual/combined_group1_con_base.rds", sep =""))
  combine = combine[order(as.numeric(row.names(combine))),] %>% as.matrix()
  path_to_Results_report = "/rds/general/project/hda_students_data/live/Group2/Alicia/Summer_project/Results_report/Multivariate_obj1/cluster_manual/group1_cen/"
  path_to_save = "/rds/general/project/hda_students_data/live/Group2/Alicia/Summer_project/Figure_report/Multivariate_obj1/cluster_manual/group1_cen/"
}

if (group2_med){
  metab_clus=readRDS(paste(path, "Results_report/cluster_manual/denoised_group2_med_base.rds", sep=""))
  metab_clus = metab_clus[order(as.numeric(row.names(metab_clus))), ] %>% as.matrix()
  combine = readRDS(paste(path, "Results_report/cluster_manual/combined_group2_med_base.rds", sep =""))
  combine = combine[order(as.numeric(row.names(combine))),] %>% as.matrix()
  path_to_Results_report = "/rds/general/project/hda_students_data/live/Group2/Alicia/Summer_project/Results_report/Multivariate_obj1/cluster_manual/group2_med/"
  path_to_save = "/rds/general/project/hda_students_data/live/Group2/Alicia/Summer_project/Figure_report/Multivariate_obj1/cluster_manual/group2_med/"
}

if (group2_cen){
  metab_clus=readRDS(paste(path, "Results_report/cluster_manual/denoised_group2_con_base.rds", sep=""))
  metab_clus = metab_clus[order(as.numeric(row.names(metab_clus))), ] %>% as.matrix()
  combine = readRDS(paste(path, "Results_report/cluster_manual/combined_group2_con_base.rds", sep =""))
  combine = combine[order(as.numeric(row.names(combine))),] %>% as.matrix()
  path_to_Results_report = "/rds/general/project/hda_students_data/live/Group2/Alicia/Summer_project/Results_report/Multivariate_obj1/cluster_manual/group2_cen/"
  path_to_save = "/rds/general/project/hda_students_data/live/Group2/Alicia/Summer_project/Figure_report/Multivariate_obj1/cluster_manual/group2_cen/"
}


if(met_no_bio){
  metab_clus = readRDS(paste(path, "Results_report/updates/denoised_clus_no_bio.rds", sep =""))
  metab_clus = metab_clus[order(as.numeric(row.names(metab_clus))), ] %>% as.matrix()
  combine = readRDS(paste(path, "Results_report/updates/combined_biometab.rds", sep=""))
  combine = combine[order(as.numeric(row.names(combine))),] %>% as.matrix()
  path_to_Results_report = "/rds/general/project/hda_students_data/live/Group2/Alicia/Summer_project/Results_report/Multivariate_obj1/Air_clus_bio/lasso/pfer/"
  path_to_save = "/rds/general/project/hda_students_data/live/Group2/Alicia/Summer_project/Figure_report/Multivariate_obj1/Air_clus_bio/lasso/pfer/"
}

if(met_base){
  biobank=readRDS(paste(path, "Results_report/biobank_final.rds", sep=""))
  metab = readRDS(paste(path, "Results_report/denoised_metab.rds", sep=""))
  metab = metab[order(as.numeric(row.names(metab))), ] %>% as.matrix()
  metab_clus = readRDS(paste(path, "Results_report/denoised_clus.rds", sep =""))
  metab_clus = metab_clus[order(as.numeric(row.names(metab_clus))), ] %>% as.matrix()
  path_to_Results_report = "/rds/general/project/hda_students_data/live/Group2/Alicia/Summer_project/Results_report/Multivariate_obj1/Air_met_base/"
  path_to_save = "/rds/general/project/hda_students_data/live/Group2/Alicia/Summer_project/Figure_report/Multivariate_obj1/Air_met_base/"
  air = select(biobank, c("24003-0.0":"24008-0.0", "24014-0.0"))
  colnames(air) <- c("NO2","NOx","PM10", "PM2.5", "PM2.5_absorbance","PM2.5-PM10", "MajorRoad")
}

if(met_first){
  biobank=readRDS(paste(path, "Results_report/matched_bb_first.rds", sep=""))
  metab = readRDS(paste(path, "Results_report/matched_metab_first.rds", sep=""))
  metab = metab %>% select(!matches("eid_52569"))
  metab = metab[order(as.numeric(row.names(metab))), ] %>% as.matrix()
  path_to_Results_report = "/rds/general/project/hda_students_data/live/Group2/Alicia/Summer_project/Results_report/Multivariate_obj1/Air_met_first/"
  path_to_save = "/rds/general/project/hda_students_data/live/Group2/Alicia/Summer_project/Figure_report/Multivariate_obj1/Air_met_first/"
  air = select(biobank, "24003-0.0":"24008-0.0")
  colnames(air) <- c("NO2","NOx","PM10", "PM2.5", "PM2.5_absorbance","PM2.5-PM10")
}

source(paste(path,"Functions/penalisation_functions.R", sep=""))
# myfunctions=list.files("/rds/general/project/hda_students_data/live/Group2/project/data_analysis/cox/functions/R")
# for (k in 1:length(myfunctions)){
#   source(paste0("/rds/general/project/hda_students_data/live/Group2/project/data_analysis/cox/functions/R/",myfunctions[k]))  
# }

#### Loading biomarker and biobank datasets ####
biobank=readRDS(paste(path, "Results_report/updates/matched_biobank.rds", sep=""))
biomarker = readRDS(paste(path, "Results_report/updates/denoised_biomarkers.rds", sep="")) %>% as.matrix()
air = select(biobank, c("24003-0.0":"24008-0.0", "24014-0.0"))
colnames(air) <- c("NO2","NOx","PM10", "PM2.5", "PM2.5_absorbance","PM2.5-PM10", "MajorRoad")


# ------------------Clustered Metabolites Stability LASSO-------------------------------------------------------
NO2 = as.matrix(air$NO2)
PM2.5 = as.matrix(air$PM2.5)
PM10 = as.matrix(air$PM10)
NOx = as.matrix(air$NOx)
PM_abs = as.matrix(air$PM2.5_absorbance)
PM2.5_10 = as.matrix(air$`PM2.5-PM10`)
air$MajorRoad = ifelse(air$MajorRoad == "No", 0,
                       ifelse(air$MajorRoad == "Yes", 1, NA ))
MajorRoad = as.matrix(air$MajorRoad)

## NO2 ##
out = VariableSelection(xdata = metab_clus, ydata = NO2,K=100, tau = 0.5, verbose=FALSE, PFER_thr =20 , PFER_method = "MB",
                          family = "gaussian")
selprop=SelectionProportions(out)

saveRDS(out, paste(path_to_Results_report, "out_metab_clus_NO2.rds", sep =""))
saveRDS(selprop, paste(path_to_Results_report, "selprop_metab_clus_NO2.rds", sep =""))

pdf(paste(path_to_save,"CalibrationPlot/clus_NO2.pdf",sep =""))
CalibrationPlot(out)
dev.off()

## PM2.5 ##
out = VariableSelection(xdata = metab_clus, ydata = PM2.5,K=100, tau = 0.5, verbose=FALSE, PFER_thr =20 , PFER_method = "MB",
                          family = "gaussian")
selprop=SelectionProportions(out)

saveRDS(out, paste(path_to_Results_report, "out_metab_clus_PM2.5.rds", sep =""))
saveRDS(selprop, paste(path_to_Results_report, "selprop_metab_clus_PM2.5.rds", sep =""))

pdf(paste(path_to_save,"CalibrationPlot/clus_PM2.5.pdf",sep =""))
CalibrationPlot(out)
dev.off()

## PM10 ##
out = VariableSelection(xdata = metab_clus, ydata = PM10,K=100, tau = 0.5, verbose=FALSE,PFER_thr =20 , PFER_method = "MB",
                          family = "gaussian")
selprop=SelectionProportions(out)

saveRDS(out, paste(path_to_Results_report, "out_metab_clus_PM10.rds", sep =""))
saveRDS(selprop, paste(path_to_Results_report, "selprop_metab_clus_PM10.rds", sep =""))

pdf(paste(path_to_save,"CalibrationPlot/clus_PM10.pdf",sep =""))
CalibrationPlot(out)
dev.off()

## NOx ##
out = VariableSelection(xdata = metab_clus, ydata = NOx,K=100, tau = 0.5, verbose=FALSE,PFER_thr =20 , PFER_method = "MB",
                          family = "gaussian")
selprop=SelectionProportions(out)

saveRDS(out, paste(path_to_Results_report, "out_metab_clus_NOx.rds", sep =""))
saveRDS(selprop, paste(path_to_Results_report, "selprop_metab_clus_NOx.rds", sep =""))

pdf(paste(path_to_save,"CalibrationPlot/clus_NOx.pdf",sep =""))
CalibrationPlot(out)
dev.off()

## PM_abs ##
out = VariableSelection(xdata = metab_clus, ydata = PM_abs,K=100, tau = 0.5, verbose=FALSE,PFER_thr =20 , PFER_method = "MB",
                        family = "gaussian")
selprop=SelectionProportions(out)

saveRDS(out, paste(path_to_Results_report, "out_metab_clus_PM_abs.rds", sep =""))
saveRDS(selprop, paste(path_to_Results_report, "selprop_metab_clus_PM_abs.rds", sep =""))

pdf(paste(path_to_save,"CalibrationPlot/clus_PM_abs.pdf",sep =""))
CalibrationPlot(out)
dev.off()

## PM2.5_10 ##
out = VariableSelection(xdata = metab_clus, ydata = PM2.5_10,K=100, tau = 0.5, verbose=FALSE,PFER_thr =20 , PFER_method = "MB",
                          family = "gaussian")
selprop=SelectionProportions(out)

saveRDS(out, paste(path_to_Results_report, "out_metab_clus_PM2.5_10.rds", sep =""))
saveRDS(selprop, paste(path_to_Results_report, "selprop_metab_clus_PM2.5_10.rds", sep =""))

pdf(paste(path_to_save,"CalibrationPlot/clus_PM2.5_10.pdf",sep =""))
CalibrationPlot(out)
dev.off()

## Major Road ##
out = VariableSelection(xdata = metab_clus, ydata = MajorRoad,K=100, tau = 0.5, verbose=FALSE,PFER_thr =20 , PFER_method = "MB",
                          family = "binomial")
selprop=SelectionProportions(out)
saveRDS(out, paste(path_to_Results_report, "out_metab_clus_road.rds", sep =""))
saveRDS(selprop, paste(path_to_Results_report, "selprop_metab_clus_road.rds", sep =""))

pdf(paste(path_to_save,"CalibrationPlot/clus_road.pdf",sep =""))
CalibrationPlot(out)
dev.off()

# ------------------Clustered Metabolites Betas parameters extraction-------------------------------------------------------
NO2 <- readRDS(paste(path_to_Results_report, "out_metab_clus_NO2.rds",sep = ""))
PM10 <- readRDS(paste(path_to_Results_report, "out_metab_clus_PM10.rds",sep = ""))
PM2.5 <- readRDS(paste(path_to_Results_report, "out_metab_clus_PM2.5.rds",sep = ""))
NOx <- readRDS(paste(path_to_Results_report, "out_metab_clus_NOx.rds",sep = ""))
PM_abs <- readRDS(paste(path_to_Results_report, "out_metab_clus_PM_abs.rds",sep = ""))
PM2.5_10 <- readRDS(paste(path_to_Results_report, "out_metab_clus_PM2.5_10.rds",sep = ""))
MajorRoad <- readRDS(paste(path_to_Results_report, "out_metab_clus_road.rds",sep = ""))

# Extracting ID of calibrated lambda
hat_lambda_id.1=ArgmaxId(NO2)[1]
hat_lambda_id.2=ArgmaxId(PM10)[1]
hat_lambda_id.3=ArgmaxId(PM2.5)[1]
hat_lambda_id.4=ArgmaxId(NOx)[1]
hat_lambda_id.5=ArgmaxId(PM_abs)[1]
hat_lambda_id.6=ArgmaxId(PM2.5_10)[1]
hat_lambda_id.7=ArgmaxId(MajorRoad)[1]

# Computing average beta coefficients from models with calibrated lambda
average_load.1=apply(NO2$Beta[hat_lambda_id.1,,],1,FUN=function(x){mean(x[x!=0])})
average_load.2=apply(PM10$Beta[hat_lambda_id.2,,],1,FUN=function(x){mean(x[x!=0])})
average_load.3=apply(PM2.5$Beta[hat_lambda_id.3,,],1,FUN=function(x){mean(x[x!=0])})
average_load.4=apply(NOx$Beta[hat_lambda_id.4,,],1,FUN=function(x){mean(x[x!=0])})
average_load.5=apply(PM_abs$Beta[hat_lambda_id.5,,],1,FUN=function(x){mean(x[x!=0])})
average_load.6=apply(PM2.5_10$Beta[hat_lambda_id.6,,],1,FUN=function(x){mean(x[x!=0])})
average_load.7=apply(MajorRoad$Beta[hat_lambda_id.7,,],1,FUN=function(x){mean(x[x!=0])})

# Save
saveRDS(average_load.1,paste(path_to_Results_report,"betas_stability_clus_NO2.rds", sep =""))
saveRDS(average_load.2,paste(path_to_Results_report,"betas_stability_clus_PM10.rds", sep =""))
saveRDS(average_load.3,paste(path_to_Results_report,"betas_stability_clus_PM2.5.rds", sep =""))
saveRDS(average_load.4,paste(path_to_Results_report,"betas_stability_clus_NOx.rds", sep =""))
saveRDS(average_load.5,paste(path_to_Results_report,"betas_stability_clus_PM_abs.rds", sep =""))
saveRDS(average_load.6,paste(path_to_Results_report,"betas_stability_clus_PM2.5_10.rds", sep =""))
saveRDS(average_load.7,paste(path_to_Results_report,"betas_stability_clus_road.rds", sep =""))


# ------Biomarkers stability LASSO -------------------------------------------------------------------

NO2 = as.matrix(air$NO2)
PM2.5 = as.matrix(air$PM2.5)
PM10 = as.matrix(air$PM10)
NOx = as.matrix(air$NOx)
PM_abs = as.matrix(air$PM2.5_absorbance)
PM2.5_10 = as.matrix(air$`PM2.5-PM10`)
MajorRoad = as.matrix(air$MajorRoad)

## NO2 ##
out = VariableSelection(xdata = biomarker, ydata = NO2,K=100, tau = 0.5, verbose=FALSE,PFER_thr =20 , PFER_method = "MB",
                        family = "gaussian")
selprop=SelectionProportions(out)

saveRDS(out, paste(path_to_Results_report, "out_bio_NO2.rds", sep =""))
saveRDS(selprop, paste(path_to_Results_report, "selprop_bio_NO2.rds", sep =""))

pdf(paste(path_to_save,"CalibrationPlot/bio_NO2.pdf",sep =""))
CalibrationPlot(out)
dev.off()

## PM2.5 ##
out = VariableSelection(xdata = biomarker, ydata = PM2.5,K=100, tau = 0.5, verbose=FALSE,PFER_thr =20 , PFER_method = "MB",
                        family = "gaussian")
selprop=SelectionProportions(out)

saveRDS(out, paste(path_to_Results_report, "out_bio_PM2.5.rds", sep =""))
saveRDS(selprop, paste(path_to_Results_report, "selprop_bio_PM2.5.rds", sep =""))

pdf(paste(path_to_save,"CalibrationPlot/bio_PM2.5.pdf",sep =""))
CalibrationPlot(out)
dev.off()

## PM10 ##
out = VariableSelection(xdata = biomarker, ydata = PM10,K=100, tau = 0.5, verbose=FALSE,PFER_thr =20 , PFER_method = "MB",
                        family = "gaussian")
selprop=SelectionProportions(out)

saveRDS(out, paste(path_to_Results_report, "out_bio_PM10.rds", sep =""))
saveRDS(selprop, paste(path_to_Results_report, "selprop_bio_PM10.rds", sep =""))

pdf(paste(path_to_save,"CalibrationPlot/bio_PM10.pdf",sep =""))
CalibrationPlot(out)
dev.off()

## NOx ##
out = VariableSelection(xdata = biomarker, ydata = NOx,K=100, tau = 0.5, verbose=FALSE,PFER_thr =20 , PFER_method = "MB",
                        family = "gaussian")
selprop=SelectionProportions(out)

saveRDS(out, paste(path_to_Results_report, "out_bio_NOx.rds", sep =""))
saveRDS(selprop, paste(path_to_Results_report, "selprop_bio_NOx.rds", sep =""))

pdf(paste(path_to_save,"CalibrationPlot/bio_NOx.pdf",sep =""))
CalibrationPlot(out)
dev.off()

## PM_abs ##
out = VariableSelection(xdata = biomarker, ydata = PM_abs,K=100, tau = 0.5, verbose=FALSE,PFER_thr =20 , PFER_method = "MB",
                        family = "gaussian")
selprop=SelectionProportions(out)

saveRDS(out, paste(path_to_Results_report, "out_bio_PM_abs.rds", sep =""))
saveRDS(selprop, paste(path_to_Results_report, "selprop_bio_PM_abs.rds", sep =""))

pdf(paste(path_to_save,"CalibrationPlot/bio_PM_abs.pdf",sep =""))
CalibrationPlot(out)
dev.off()

## PM2.5_10 ##
out = VariableSelection(xdata = biomarker, ydata = PM2.5_10,K=100, tau = 0.5, verbose=FALSE,PFER_thr =20 , PFER_method = "MB",
                        family = "gaussian")
selprop=SelectionProportions(out)

saveRDS(out, paste(path_to_Results_report, "out_bio_PM2.5_10.rds", sep =""))
saveRDS(selprop, paste(path_to_Results_report, "selprop_bio_PM2.5_10.rds", sep =""))

pdf(paste(path_to_save,"CalibrationPlot/bio_PM2.5_10.pdf",sep =""))
CalibrationPlot(out)
dev.off()

## Major Road ##
out = VariableSelection(xdata = biomarker, ydata = MajorRoad,K=100, tau = 0.5, verbose=FALSE,PFER_thr =20 , PFER_method = "MB",
                        family = "binomial")
selprop=SelectionProportions(out)

saveRDS(out, paste(path_to_Results_report, "out_bio_road.rds", sep =""))
saveRDS(selprop, paste(path_to_Results_report, "selprop_bio_road.rds", sep =""))

pdf(paste(path_to_save,"CalibrationPlot/bio_road.pdf",sep =""))
CalibrationPlot(out)
dev.off()

# ------------------Biomarkers Betas parameters extraction-------------------------------------------------------

NO2 <- readRDS(paste(path_to_Results_report, "out_bio_NO2.rds",sep = ""))
PM10 <- readRDS(paste(path_to_Results_report, "out_bio_PM10.rds",sep = ""))
PM2.5 <- readRDS(paste(path_to_Results_report, "out_bio_PM2.5.rds",sep = ""))
NOx <- readRDS(paste(path_to_Results_report, "out_bio_NOx.rds",sep = ""))
PM_abs <- readRDS(paste(path_to_Results_report, "out_bio_PM_abs.rds",sep = ""))
PM2.5_10 <- readRDS(paste(path_to_Results_report, "out_bio_PM2.5_10.rds",sep = ""))
MajorRoad <- readRDS(paste(path_to_Results_report, "out_bio_road.rds",sep = ""))

# Extracting ID of calibrated lambda
hat_lambda_id.1=ArgmaxId(NO2)[1]
hat_lambda_id.2=ArgmaxId(PM10)[1]
hat_lambda_id.3=ArgmaxId(PM2.5)[1]
hat_lambda_id.4=ArgmaxId(NOx)[1]
hat_lambda_id.5=ArgmaxId(PM_abs)[1]
hat_lambda_id.6=ArgmaxId(PM2.5_10)[1]
hat_lambda_id.7=ArgmaxId(MajorRoad)[1]

# Computing average beta coefficients from models with calibrated lambda
average_load.1=apply(NO2$Beta[hat_lambda_id.1,,],1,FUN=function(x){mean(x[x!=0])})
average_load.2=apply(PM10$Beta[hat_lambda_id.2,,],1,FUN=function(x){mean(x[x!=0])})
average_load.3=apply(PM2.5$Beta[hat_lambda_id.3,,],1,FUN=function(x){mean(x[x!=0])})
average_load.4=apply(NOx$Beta[hat_lambda_id.4,,],1,FUN=function(x){mean(x[x!=0])})
average_load.5=apply(PM_abs$Beta[hat_lambda_id.5,,],1,FUN=function(x){mean(x[x!=0])})
average_load.6=apply(PM2.5_10$Beta[hat_lambda_id.6,,],1,FUN=function(x){mean(x[x!=0])})
average_load.7=apply(MajorRoad$Beta[hat_lambda_id.7,,],1,FUN=function(x){mean(x[x!=0])})

# Save
saveRDS(average_load.1,paste(path_to_Results_report,"betas_bio_NO2.rds", sep =""))
saveRDS(average_load.2,paste(path_to_Results_report,"betas_bio_PM10.rds", sep =""))
saveRDS(average_load.3,paste(path_to_Results_report,"betas_bio_PM2.5.rds", sep =""))
saveRDS(average_load.4,paste(path_to_Results_report,"betas_bio_NOx.rds", sep =""))
saveRDS(average_load.5,paste(path_to_Results_report,"betas_bio_PM_abs.rds", sep =""))
saveRDS(average_load.6,paste(path_to_Results_report,"betas_bio_PM2.5_10.rds", sep =""))
saveRDS(average_load.7,paste(path_to_Results_report,"betas_bio_road.rds", sep =""))


# ------Combined stability LASSO -------------------------------------------------------------------

NO2 = as.matrix(air$NO2)
PM2.5 = as.matrix(air$PM2.5)
PM10 = as.matrix(air$PM10)
NOx = as.matrix(air$NOx)
PM_abs = as.matrix(air$PM2.5_absorbance)
PM2.5_10 = as.matrix(air$`PM2.5-PM10`)
MajorRoad = as.matrix(air$MajorRoad)

## NO2 ##
out = VariableSelection(xdata = combine, ydata = NO2,K=100, tau = 0.5, verbose=FALSE,PFER_thr =20 , PFER_method = "MB",
                        family = "gaussian")
selprop=SelectionProportions(out)

saveRDS(out, paste(path_to_Results_report, "out_combine_NO2.rds", sep =""))
saveRDS(selprop, paste(path_to_Results_report, "selprop_combine_NO2.rds", sep =""))

pdf(paste(path_to_save,"CalibrationPlot/combine_NO2.pdf",sep =""))
CalibrationPlot(out)
dev.off()

## PM2.5 ##
out = VariableSelection(xdata = combine, ydata = PM2.5,K=100, tau = 0.5, verbose=FALSE,PFER_thr =20 , PFER_method = "MB",
                        family = "gaussian")
selprop=SelectionProportions(out)

saveRDS(out, paste(path_to_Results_report, "out_combine_PM2.5.rds", sep =""))
saveRDS(selprop, paste(path_to_Results_report, "selprop_combine_PM2.5.rds", sep =""))

pdf(paste(path_to_save,"CalibrationPlot/combine_PM2.5.pdf",sep =""))
CalibrationPlot(out)
dev.off()

## PM10 ##
out = VariableSelection(xdata = combine, ydata = PM10,K=100, tau = 0.5, verbose=FALSE,PFER_thr =20 , PFER_method = "MB",
                        family = "gaussian")
selprop=SelectionProportions(out)

saveRDS(out, paste(path_to_Results_report, "out_combine_PM10.rds", sep =""))
saveRDS(selprop, paste(path_to_Results_report, "selprop_combine_PM10.rds", sep =""))

pdf(paste(path_to_save,"CalibrationPlot/combine_PM10.pdf",sep =""))
CalibrationPlot(out)
dev.off()

## NOx ##
out = VariableSelection(xdata = combine, ydata = NOx,K=100, tau = 0.5, verbose=FALSE,PFER_thr =20 , PFER_method = "MB",
                        family = "gaussian")
selprop=SelectionProportions(out)

saveRDS(out, paste(path_to_Results_report, "out_combine_NOx.rds", sep =""))
saveRDS(selprop, paste(path_to_Results_report, "selprop_combine_NOx.rds", sep =""))

pdf(paste(path_to_save,"CalibrationPlot/combine_NOx.pdf",sep =""))
CalibrationPlot(out)
dev.off()

## PM_abs ##
out = VariableSelection(xdata = combine, ydata = PM_abs,K=100, tau = 0.5, verbose=FALSE,PFER_thr =20 , PFER_method = "MB",
                        family = "gaussian")
selprop=SelectionProportions(out)

saveRDS(out, paste(path_to_Results_report, "out_combine_PM_abs.rds", sep =""))
saveRDS(selprop, paste(path_to_Results_report, "selprop_combine_PM_abs.rds", sep =""))

pdf(paste(path_to_save,"CalibrationPlot/combine_PM_abs.pdf",sep =""))
CalibrationPlot(out)
dev.off()

## PM2.5_10 ##
out = VariableSelection(xdata = combine, ydata = PM2.5_10,K=100, tau = 0.5, verbose=FALSE,PFER_thr =20 , PFER_method = "MB",
                        family = "gaussian")
selprop=SelectionProportions(out)

saveRDS(out, paste(path_to_Results_report, "out_combine_PM2.5_10.rds", sep =""))
saveRDS(selprop, paste(path_to_Results_report, "selprop_combine_PM2.5_10.rds", sep =""))

pdf(paste(path_to_save,"CalibrationPlot/combine_PM2.5_10.pdf",sep =""))
CalibrationPlot(out)
dev.off()

## Major Road ##
out = VariableSelection(xdata = combine, ydata = MajorRoad,K=100, tau = 0.5, verbose=FALSE,PFER_thr =20 , PFER_method = "MB",
                        family = "binomial")
selprop=SelectionProportions(out)

saveRDS(out, paste(path_to_Results_report, "out_combine_road.rds", sep =""))
saveRDS(selprop, paste(path_to_Results_report, "selprop_combine_road.rds", sep =""))

pdf(paste(path_to_save,"CalibrationPlot/combine_road.pdf",sep =""))
CalibrationPlot(out)
dev.off()

# ------------------Combined Betas parameters extraction-------------------------------------------------------
NO2 <- readRDS(paste(path_to_Results_report, "out_combine_NO2.rds",sep = ""))
PM10 <- readRDS(paste(path_to_Results_report, "out_combine_PM10.rds",sep = ""))
PM2.5 <- readRDS(paste(path_to_Results_report, "out_combine_PM2.5.rds",sep = ""))
NOx <- readRDS(paste(path_to_Results_report, "out_combine_NOx.rds",sep = ""))
PM_abs <- readRDS(paste(path_to_Results_report, "out_combine_PM_abs.rds",sep = ""))
PM2.5_10 <- readRDS(paste(path_to_Results_report, "out_combine_PM2.5_10.rds",sep = ""))
MajorRoad <- readRDS(paste(path_to_Results_report, "out_combine_road.rds",sep = ""))

# # Extracting ID of calibrated lambda
hat_lambda_id.1=ArgmaxId(NO2)[1]
hat_lambda_id.2=ArgmaxId(PM10)[1]
hat_lambda_id.3=ArgmaxId(PM2.5)[1]
hat_lambda_id.4=ArgmaxId(NOx)[1]
hat_lambda_id.5=ArgmaxId(PM_abs)[1]
hat_lambda_id.6=ArgmaxId(PM2.5_10)[1]
hat_lambda_id.7=ArgmaxId(MajorRoad)[1]

# # Computing average beta coefficients from models with calibrated lambda
average_load.1=apply(NO2$Beta[hat_lambda_id.1,,],1,FUN=function(x){mean(x[x!=0])})
average_load.2=apply(PM10$Beta[hat_lambda_id.2,,],1,FUN=function(x){mean(x[x!=0])})
average_load.3=apply(PM2.5$Beta[hat_lambda_id.3,,],1,FUN=function(x){mean(x[x!=0])})
average_load.4=apply(NOx$Beta[hat_lambda_id.4,,],1,FUN=function(x){mean(x[x!=0])})
average_load.5=apply(PM_abs$Beta[hat_lambda_id.5,,],1,FUN=function(x){mean(x[x!=0])})
average_load.6=apply(PM2.5_10$Beta[hat_lambda_id.6,,],1,FUN=function(x){mean(x[x!=0])})
average_load.7=apply(MajorRoad$Beta[hat_lambda_id.7,,],1,FUN=function(x){mean(x[x!=0])})

# # Save
saveRDS(average_load.1,paste(path_to_Results_report,"betas_combine_NO2.rds", sep =""))
saveRDS(average_load.2,paste(path_to_Results_report,"betas_combine_PM10.rds", sep =""))
saveRDS(average_load.3,paste(path_to_Results_report,"betas_combine_PM2.5.rds", sep =""))
saveRDS(average_load.4,paste(path_to_Results_report,"betas_combine_NOx.rds", sep =""))
saveRDS(average_load.5,paste(path_to_Results_report,"betas_combine_PM_abs.rds", sep =""))
saveRDS(average_load.6,paste(path_to_Results_report,"betas_combine_PM2.5_10.rds", sep =""))
saveRDS(average_load.7,paste(path_to_Results_report,"betas_combine_road.rds", sep =""))



# # --------------------------Individual Metabolites Stability LASSO-----------------------------------------------
# NO2 = as.matrix(air$NO2)
# PM2.5 = as.matrix(air$PM2.5)
# PM10 = as.matrix(air$PM10)
# NOx = as.matrix(air$NOx)
# PM_abs = as.matrix(air$PM2.5_absorbance)
# PM2.5_10 = as.matrix(air$`PM2.5-PM10`)
# air$MajorRoad = ifelse(air$MajorRoad == "No", 0,
#                        ifelse(air$MajorRoad == "Yes", 1, NA ))
# MajorRoad = as.matrix(air$MajorRoad)
# 
# ## NO2 ##
# out = VariableSelection(xdata = metab, ydata = NO2,K=100, tau = 0.5, verbose=FALSE,
#                            family = "gaussian")
# selprop=SelectionProportions(out)
# 
# saveRDS(out, paste(path_to_Results_report, "out_metab_NO2.rds", sep =""))
# saveRDS(selprop, paste(path_to_Results_report, "selprop_metab_NO2.rds", sep =""))
# 
# pdf(paste(path_to_save,"CalibrationPlot/NO2.pdf",sep =""))
# CalibrationPlot(out)
# dev.off()
# 
# ## PM2.5 ##
# out = VariableSelection(xdata = metab, ydata = PM2.5,K=100, tau = 0.5, verbose=FALSE,
#                           family = "gaussian")
# selprop=SelectionProportions(out)
# 
# saveRDS(out, paste(path_to_Results_report, "out_metab_PM2.5.rds", sep =""))
# saveRDS(selprop, paste(path_to_Results_report, "selprop_metab_PM2.5.rds", sep =""))
# 
# pdf(paste(path_to_save,"CalibrationPlot/PM2.5.pdf",sep =""))
# CalibrationPlot(out)
# dev.off()
# 
# ## PM10 ##
# out = VariableSelection(xdata = metab, ydata = PM10,K=100, tau = 0.5, verbose=FALSE,
#                           family = "gaussian")
# selprop=SelectionProportions(out)
# 
# saveRDS(out, paste(path_to_Results_report, "out_metab_PM10.rds", sep =""))
# saveRDS(selprop, paste(path_to_Results_report, "selprop_metab_PM10.rds", sep =""))
# 
# pdf(paste(path_to_save,"CalibrationPlot/PM10.pdf",sep =""))
# CalibrationPlot(out)
# dev.off()
# 
# ## NOx ##
# out = VariableSelection(xdata = metab, ydata = NOx,K=100, tau = 0.5, verbose=FALSE,
#                           family = "gaussian")
# selprop=SelectionProportions(out)
# 
# saveRDS(out, paste(path_to_Results_report, "out_metab_NOx.rds", sep =""))
# saveRDS(selprop, paste(path_to_Results_report, "selprop_metab_NOx.rds", sep =""))
# 
# pdf(paste(path_to_save,"CalibrationPlot/NOx.pdf",sep =""))
# CalibrationPlot(out)
# dev.off()
# 
# ## PM_abs ##
# out = VariableSelection(xdata = metab, ydata = PM_abs,K=100, tau = 0.5, verbose=FALSE,
#                           family = "gaussian")
# selprop=SelectionProportions(out)
# 
# saveRDS(out, paste(path_to_Results_report, "out_metab_PM_abs.rds", sep =""))
# saveRDS(selprop, paste(path_to_Results_report, "selprop_metab_PM_abs.rds", sep =""))
# 
# pdf(paste(path_to_save,"CalibrationPlot/PM_abs.pdf",sep =""))
# CalibrationPlot(out)
# dev.off()
# 
# ## PM2.5_10 ##
# out = VariableSelection(xdata = metab, ydata = PM2.5_10,K=100, tau = 0.5, verbose=FALSE,
#                           family = "gaussian")
# selprop=SelectionProportions(out)
# 
# saveRDS(out, paste(path_to_Results_report, "out_metab_PM2.5_10.rds", sep =""))
# saveRDS(selprop, paste(path_to_Results_report, "selprop_metab_PM2.5_10.rds", sep =""))
# 
# pdf(paste(path_to_save,"CalibrationPlot/PM2.5_10.pdf",sep =""))
# CalibrationPlot(out)
# dev.off()
# 
# ## Major Road ##
# out = VariableSelection(xdata = metab, ydata = MajorRoad,K=100, tau = 0.5, verbose=FALSE,
#                           family = "gaussian")
# selprop=SelectionProportions(out)
# 
# saveRDS(out, paste(path_to_Results_report, "out_metab_road.rds", sep =""))
# saveRDS(selprop, paste(path_to_Results_report, "selprop_metab_road.rds", sep =""))
# 
# pdf(paste(path_to_save,"CalibrationPlot/road.pdf",sep =""))
# CalibrationPlot(out)
# dev.off()
# 
# # Multivariate responses ##
# # all = as.matrix(cbind(NO2, PM2.5, PM10))
# # 
# # out = VariableSelection(xdata = metab, ydata = all,K=100, Lambda = 1:ncol(metab),
# #                           family = "mgaussian")
# # selprop=SelectionProportions(out, plot=FALSE)
# # 
# # saveRDS(out, paste(path_to_Results_report, "out_metab_multi.rds", sep =""))
# # saveRDS(selprop, paste(path_to_Results_report, "selprop_metab_multi.rds", sep =""))
# 
# 
# 
# # ------------------Individual Metabolites Betas parameters extraction-------------------------------------------------------
# NO2 <- readRDS(paste(path_to_Results_report, "out_metab_NO2.rds",sep = ""))
# PM10 <- readRDS(paste(path_to_Results_report, "out_metab_PM10.rds",sep = ""))
# PM2.5 <- readRDS(paste(path_to_Results_report, "out_metab_PM2.5.rds",sep = ""))
# NOx <- readRDS(paste(path_to_Results_report, "out_metab_NOx.rds",sep = ""))
# PM_abs <- readRDS(paste(path_to_Results_report, "out_metab_PM_abs.rds",sep = ""))
# PM2.5_10 <- readRDS(paste(path_to_Results_report, "out_metab_PM2.5_10.rds",sep = ""))
# MajorRoad <- readRDS(paste(path_to_Results_report, "out_metab_road.rds",sep = ""))
# 
# # multi <- readRDS(paste(path_to_Results_report, "out_metab_multi.rds",sep = ""))
# 
# # Extracting ID of calibrated lambda
# hat_lambda_id.1=ArgmaxId(NO2)[1]
# hat_lambda_id.2=ArgmaxId(PM10)[1]
# hat_lambda_id.3=ArgmaxId(PM2.5)[1]
# hat_lambda_id.4=ArgmaxId(NOx)[1]
# hat_lambda_id.5=ArgmaxId(PM_abs)[1]
# hat_lambda_id.6=ArgmaxId(PM2.5_10)[1]
# hat_lambda_id.7=ArgmaxId(MajorRoad)[1]
# # hat_lambda_id.4=ArgmaxId(multi)[1]
# 
# # Computing average beta coefficients from models with calibrated lambda
# average_load.1=apply(NO2$Beta[hat_lambda_id.1,,],1,FUN=function(x){mean(x[x!=0])})
# average_load.2=apply(PM10$Beta[hat_lambda_id.2,,],1,FUN=function(x){mean(x[x!=0])})
# average_load.3=apply(PM2.5$Beta[hat_lambda_id.3,,],1,FUN=function(x){mean(x[x!=0])})
# average_load.4=apply(NOx$Beta[hat_lambda_id.4,,],1,FUN=function(x){mean(x[x!=0])})
# average_load.5=apply(PM_abs$Beta[hat_lambda_id.5,,],1,FUN=function(x){mean(x[x!=0])})
# average_load.6=apply(PM2.5_10$Beta[hat_lambda_id.6,,],1,FUN=function(x){mean(x[x!=0])})
# average_load.7=apply(MajorRoad$Beta[hat_lambda_id.7,,],1,FUN=function(x){mean(x[x!=0])})
# # average_load.4=apply(multi$Beta[hat_lambda_id.4,,],1,FUN=function(x){mean(x[x!=0])})
# 
# # Save
# saveRDS(average_load.1,paste(path_to_Results_report,"betas_stability_NO2.rds", sep =""))
# saveRDS(average_load.2,paste(path_to_Results_report,"betas_stability_PM10.rds", sep =""))
# saveRDS(average_load.3,paste(path_to_Results_report,"betas_stability_PM2.5.rds", sep =""))
# saveRDS(average_load.4,paste(path_to_Results_report,"betas_stability_NOx.rds", sep =""))
# saveRDS(average_load.5,paste(path_to_Results_report,"betas_stability_PM_abs.rds", sep =""))
# saveRDS(average_load.6,paste(path_to_Results_report,"betas_stability_PM2.5_10.rds", sep =""))
# saveRDS(average_load.7,paste(path_to_Results_report,"betas_stability_road.rds", sep =""))
